requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.11.0=py37h89c1867_0
apptools=4.5.0=pyh9f0ad1d_1
argon2-cffi=20.1.0=py37h4abf009_2
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.2.1=pyh9f0ad1d_0
blessings=1.7=pypi_0
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h516909a_3
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cairo=1.16.0=h488836b_1006
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.4=py37h11fe52a_0
chardet=4.0.0=py37h89c1867_0
configobj=5.0.6=py_0
cryptography=3.3.1=py37h3c74f83_0
cudatoolkit=10.2.89=hfd86e86_1
curl=7.71.1=he644dc0_8
cycler=0.10.0=py_2
dbus=1.13.6=hfdff14a_1
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
double-conversion=3.1.5=he1b5a44_2
eigen=3.3.8=h0efe328_0
entrypoints=0.3=pyhd8ed1ab_1003
envisage=4.9.2=pyh9f0ad1d_1
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h3215721_1
fontconfig=2.13.1=h7e3eb15_1002
freeglut=3.2.1=h58526e2_0
freetype=2.10.4=h5ab3b9f_0
gettext=0.19.8.1=hf34092f_1004
gl2ps=1.4.2=h0708190_0
glew=2.1.0=he1b5a44_0
glib=2.66.3=h58526e2_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gpustat=0.6.0=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.33.2=py37haffed2e_2
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.7.2=ha5b49bf_1
hdf4=4.2.13=hf30be14_1003
hdf5=1.10.6=nompi_h7c3c948_1111
icu=67.1=he1b5a44_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.3.0=py37h89c1867_2
importlib_metadata=3.3.0=hd8ed1ab_2
intel-openmp=2020.2=254
ipykernel=5.4.2=py37h888b3d9_0
ipython=5.8.0=py37_1
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jasper=1.900.1=h07fcdf6_1006
jinja2=2.11.2=pyh9f0ad1d_0
jpeg=9d=h36c2ea0_0
json5=0.9.5=pyh9f0ad1d_0
jsoncpp=1.8.4=hc9558a2_1002
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=5.2.0=py37_1
jupyter_core=4.7.0=py37h89c1867_0
jupyterlab=2.2.9=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py37hc928c03_0
krb5=1.17.2=h926e7f8_0
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.9.0=1_h6e990d7_netlib
libcblas=3.9.0=3_h893e4fe_netlib
libclang=10.0.1=default_hde54327_1
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=h14c3975_1
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran=3.0.0=1
libgfortran-ng=7.5.0=hae1eefd_17
libgfortran4=7.5.0=hae1eefd_17
libglib=2.66.3=hbe7bbb4_0
libglu=9.0.0=he1b5a44_1001
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
liblapack=3.9.0=3_h893e4fe_netlib
liblapacke=3.9.0=3_h893e4fe_netlib
libllvm10=10.0.1=he513fc3_3
libnetcdf=4.7.4=nompi_h56d31a8_107
libnghttp2=1.41.0=h8cfc5f6_2
libogg=1.3.4=h7f98852_0
libopencv=4.5.0=py37_3
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.3=h255efa7_3
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.1.0=hdf63c60_0
libtheora=1.1.1=h516909a_1004
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=he1b5a44_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.1.0=h36c2ea0_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=h68273f3_2
loguru=0.5.3=py37hc8dfbb8_2
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py37hb5d75c8_2
matplotlib-base=3.3.3=py37h4f6019d_0
mayavi=4.7.2=py37hdf92800_4
mistune=0.8.4=py37h4abf009_1002
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
mysql-common=8.0.21=2
mysql-libs=8.0.21=hf3661c5_2
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37h89c1867_3
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
ninja=1.10.2=py37hff7bd54_0
notebook=6.1.5=py37h89c1867_0
nspr=4.29=he1b5a44_1
nss=3.59=h2c00c37_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
nvidia-ml-py3=7.352.0=pypi_0
olefile=0.46=py37_0
opencv=4.5.0=py37_3
openh264=2.1.1=h8b12597_0
openssl=1.1.1i=h7f98852_0
packaging=20.8=pyhd3deb0d_0
pandas=1.1.4=py37h10a2094_0
pandoc=2.11.3.1=h7f98852_0
pandocfilters=1.4.2=py_1
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.1=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
pixman=0.38.0=h516909a_1003
proj=7.1.1=h966b41f_3
prometheus_client=0.9.0=pyhd3deb0d_0
prompt_toolkit=1.0.15=py_1
protobuf=3.14.0=py37h2531618_1
psutil=5.8.0=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.6.0=py_1001
pugixml=1.10=he1b5a44_1
py-opencv=4.5.0=py37h888b3d9_3
pycparser=2.20=pyh9f0ad1d_2
pyface=7.1.0=pyh9f0ad1d_0
pygments=2.7.3=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h8685d9f_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=py37h4abf009_1
pysocks=1.7.1=py37he5f6b98_2
python=3.7.8=h6f2ec95_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2020.5=pyhd8ed1ab_0
pyyaml=5.3.1=py37h5e8e339_2
pyzmq=20.0.0=py37h5a562af_1
qt=5.12.9=h1f2b2cb_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.25.1=pyhd3deb0d_0
scipy=1.5.3=py37h8911b10_0
seaborn=0.11.1=ha770c72_0
seaborn-base=0.11.1=pyhd8ed1ab_0
send2trash=1.5.0=py_0
setuptools=51.0.0=py37h06a4308_2
simplegeneric=0.8.1=py_1
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
statsmodels=0.12.1=py37h902c9e0_2
tbb=2020.2=hc9558a2_0
tbb-devel=2020.2=hc9558a2_0
tensorboard=1.15.0=py37_0
terminado=0.9.1=py37h89c1867_1
testpath=0.4.4=py_0
tk=8.6.10=hbc83047_0
torchaudio=0.7.2=py37
torchvision=0.2.2=py_3
tornado=6.1=py37h4abf009_0
tqdm=4.56.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
traits=6.1.1=py37h4abf009_0
traitsui=7.1.0=pyh9f0ad1d_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pyhd8ed1ab_0
utfcpp=3.1.2=0
vtk=9.0.1=no_osmesa_py37h8af0cc8_102
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
x264=1!152.20180806=h14c3975_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-libxt=1.2.0=h516909a_0
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zeromq=4.3.3=h58526e2_3
zipp=3.4.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
